rearrange vhci pnp related codes
Old vhci was mixed up with framework routines required by pnp manager
and device(vhub/vpdo) manipulation routines. Thus, this commit
rearranges codes as follows:
- vhci_pnp.c: common pnp routines(vhub/vpdo)
- vhci_pnp_vhub: vhub pnp routines(newly created)
- vhci_pnp_vpdo: vpdo pnp routines(newly created)
- vhci_vhub:  vhub manipulation routines(newly created)
- vhci_vpdo:  vpdo manipulation routines

Only code relocation and beautify without new feature are done.
